refactor(wrapper): move CodexProvider to its own file and restore technical comments
