+++ packages/wrapper/src/cli.ts
@@ -23,11 +23,7 @@ const execFileAsync = promisify(execFile);
 
 const VERSION = loadVersion();
 const EXIT_ERROR = 1;
-const VALID_PERMISSION_PROFILES: PermissionProfile[] = [
-  'default',
-  'restricted',
-  'unrestricted',
-];
+const VALID_PERMISSION_PROFILES: PermissionProfile[] = ['default', 'restricted', 'unrestricted'];
 
 async function main(): Promise<void> {
   const [, , subcommand, ...rest] = process.argv;
@@ -125,6 +121,9 @@ async function cmdProvider(args: string[]): Promise<void> {
   await providerSetup(providerName);
 }
 
+// ---------------------------------------------------------------------------
+// aco run <provider> <command> [--input <text>] [--permission-profile <profile>]
+// ---------------------------------------------------------------------------
 async function cmdRun(args: string[]): Promise<void> {
   if (args.includes('--help') || args.includes('-h')) {
     console.error(
@@ -149,8 +148,7 @@ async function cmdRun(args: string[]): Promise<void> {
     process.exit(EXIT_ERROR);
   }
 
-  const permissionProfile =
-    parseFlag<PermissionProfile>(args, '--permission-profile') ?? 'default';
+  const permissionProfile = parseFlag<PermissionProfile>(args, '--permission-profile') ?? 'default';
   if (!VALID_PERMISSION_PROFILES.includes(permissionProfile)) {
     console.error(
       `Invalid --permission-profile '${permissionProfile}'. Valid values: default|restricted|unrestricted`
@@ -189,12 +187,7 @@ async function cmdRun(args: string[]): Promise<void> {
     prompt = await readFile(globalPromptPath, 'utf8');
   }
 
-  const session = await sessionStore.create(
-    providerKey,
-    command,
-    undefined,
-    permissionProfile
-  );
+  const session = await sessionStore.create(providerKey, command, undefined, permissionProfile);
   const tee = sessionStore.createOutputTee(session.id);
   let hasOutput = false;
   let runError: unknown;
@@ -204,6 +197,7 @@ async function cmdRun(args: string[]): Promise<void> {
       permissionProfile,
       sessionId: session.id,
       onPid: (pid) => {
+        // Fire-and-forget pid update so the session can be cancelled
         sessionStore.update(session.id, { pid }).catch((err: unknown) => {
           console.warn(
             'Failed to record process PID:',
@@ -308,7 +302,7 @@ async function cmdCancel(args: string[]): Promise<void> {
     try {
       process.kill(record.pid, 'SIGTERM');
     } catch {
-      // The process may have already exited.
+      // The process may have already exited — safe to ignore
     }
   }
